add glm as external library

+++ src/VoxUpdate.cpp
@@ -220,7 +220,7 @@ void VoxGame::Update()
 
 void VoxGame::UpdateNamePicking()
 {
-	POINT lMouse = { VoxGame::GetInstance()->GetWindowCursorX(), (m_windowHeight - VoxGame::GetInstance()->GetWindowCursorY()) };
+	POINT lMouse = { (float)VoxGame::GetInstance()->GetWindowCursorX(), (float)(m_windowHeight - VoxGame::GetInstance()->GetWindowCursorY()) };
 
 	// Push attribs
 	glPushAttrib(GL_LIST_BIT | GL_CURRENT_BIT | GL_ENABLE_BIT | GL_TRANSFORM_BIT);
